requirements.txt
# Agent dependencies
psutil==5.9.5
requests==2.32.3
python-dotenv==1.0.0

# Server dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
gunicorn==21.2.0
